Add (customizable) PreConditioner method in Problem
Similarly to Manopt interface, we include PreConditioner as a method
in Problem (defaulted to identity map) that can be overloaded by
the user for implementing specific PreConditioner for a problem.
This follows the ROPTLIB philosophy clearly dividing solver and problem.

The Problem PreConditioner is applied via RTRNewton::PreConditioner,
which specializes wrt SolversTR::PreConditioner.