Refactor some symbols to avoid conflicts with other libraries
1.
Fortran wrappers define multiple types and macros
in the global namespace that conflict e.g. cmath, Eigen, etc.
To minimize these conflicts (which depend on #include ordering
and are thus very hard to detect and fix):
- We leave one single f2c.h file common to blas and lapack
  Currently, both have the same content and include guard,
  so keeping both is a source for conflicts.
- In f2c.h, we redefine real as f2c_real to avoid conflict with Eigen
- In f2c.h, unused macros are commented out

2. In ROPTLIB, Variable and Vector are just Element.
So far this was implemented by
```
```
but the use of #define has no clear scope and might interfere with other
libraries which use the keyword Vector or Variable for other purposes.
Because of this, it is equivalent but much safer to define Variable and Vector
as aliases to the same Element class.
We add these `typedef`'s in the base file *Element.h*.